pdf: Avoid mutating reportlab constants
`GRID_STYLE` is mutated every time we generate a PDF. Instead of mutating
it, we create a new `TableStyle`, which inherits from `GRID_STYLE`, on
each request.

When deployed on mod_wsgi, we usually had a new clean Python process for
each PDF we generated. After we moved to uwsgi, the Python process is
reused for many requests, and this bug becomes much more visible.

Credits to Thomas Adamcik for figuring out the reason for the PDF
generation crashes.